Refactor : Classes needed to move.
